requirements.txt
albumentations>=1.4.0
numpy>=1.23
pandas>=2.0
pyyaml>=6.0
scikit-learn>=1.3
timm>=0.9.0
torch>=2.1
torchvision>=0.16
